non-affine hexadron working
