+++ cpp/kernels.hpp
@@ -22,6 +22,7 @@ enum Kernel
 {
   Mass,
   MassTensor,
+  MassNonAffine,
   Stiffness,
   SymGrad
 };
@@ -77,9 +78,10 @@ kernel_fn generate_tet_kernel(dolfinx_cuas::Kernel type)
       for (std::int32_t i = 0; i < ndofs_cell; i++)
         _dphi(q, i, k) = dphi(k, q, i);
 
-  kernel_fn stiffness = [=](double* A, const double* c, const double* w,
-                            const double* coordinate_dofs, const int* entity_local_index,
-                            const std::uint8_t* quadrature_permutation) {
+  kernel_fn stiffness
+      = [=](double* A, const double* c, const double* w, const double* coordinate_dofs,
+            const int* entity_local_index, const std::uint8_t* quadrature_permutation)
+  {
     // Get geometrical data
     xt::xtensor<double, 2> J = xt::zeros<double>({gdim, tdim});
     xt::xtensor<double, 2> K = xt::zeros<double>({tdim, gdim});
@@ -122,7 +124,8 @@ kernel_fn generate_tet_kernel(dolfinx_cuas::Kernel type)
   // Mass Matrix using quadrature formulation
   // =====================================================================================
   kernel_fn mass = [=](double* A, const double* c, const double* w, const double* coordinate_dofs,
-                       const int* entity_local_index, const std::uint8_t* quadrature_permutation) {
+                       const int* entity_local_index, const std::uint8_t* quadrature_permutation)
+  {
     // Get geometrical data
     xt::xtensor<double, 2> J = xt::zeros<double>({gdim, tdim});
     std::array<std::size_t, 2> shape = {d, gdim};
@@ -155,9 +158,10 @@ kernel_fn generate_tet_kernel(dolfinx_cuas::Kernel type)
       for (int j = 0; j < ndofs_cell; j++)
         A0(i, j) += weights[q] * phi(q, i) * phi(q, j);
 
-  kernel_fn masstensor = [=](double* A, const double* c, const double* w,
-                              const double* coordinate_dofs, const int* entity_local_index,
-                              const std::uint8_t* quadrature_permutation) {
+  kernel_fn masstensor
+      = [=](double* A, const double* c, const double* w, const double* coordinate_dofs,
+            const int* entity_local_index, const std::uint8_t* quadrature_permutation)
+  {
     // Get geometrical data
     xt::xtensor<double, 2> J = xt::zeros<double>({gdim, tdim});
     std::array<std::size_t, 2> shape = {d, gdim};
